refactor(runtime,microbenchmarks): fix split package io.kroxylicious.proxy.filter
Move filter invoker machinery (FilterInvoker, FilterInvokers,
FilterAndInvoker, SafeInvoker, and the per-message invoker classes)
from io.kroxylicious.proxy.filter to io.kroxylicious.proxy.internal.filter
in the runtime module, making previously package-private classes public.

Move internal Filter implementations (BrokerAddressFilter,
ApiVersionsDowngradeFilter, ApiVersionsIntersectFilter,
EagerMetadataLearner, TopicNameCacheFilter) to
io.kroxylicious.proxy.internal.filter.impl to keep the machinery
and implementation concerns separate.

Move microbenchmark invoker classes (ArrayFilterInvoker,
SpecificFilterInvoker) to io.kroxylicious.benchmarking.jmh,
eliminating the split package with kroxylicious-api.

Update FreeMarker templates to generate into the new package
with public visibility.
