+++ kroxylicious-microbenchmarks/src/main/java/io/kroxylicious/benchmarking/jmh/ArrayFilterInvoker.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.kroxylicious.proxy.filter;
+package io.kroxylicious.benchmarking.jmh;
 
 import java.util.Arrays;
 import java.util.HashMap;
@@ -16,6 +16,269 @@ import org.apache.kafka.common.message.RequestHeaderData;
 import org.apache.kafka.common.message.ResponseHeaderData;
 import org.apache.kafka.common.protocol.ApiKeys;
 import org.apache.kafka.common.protocol.ApiMessage;
+
+import io.kroxylicious.proxy.filter.AddOffsetsToTxnRequestFilter;
+import io.kroxylicious.proxy.filter.AddOffsetsToTxnResponseFilter;
+import io.kroxylicious.proxy.filter.AddPartitionsToTxnRequestFilter;
+import io.kroxylicious.proxy.filter.AddPartitionsToTxnResponseFilter;
+import io.kroxylicious.proxy.filter.AllocateProducerIdsRequestFilter;
+import io.kroxylicious.proxy.filter.AllocateProducerIdsResponseFilter;
+import io.kroxylicious.proxy.filter.AlterClientQuotasRequestFilter;
+import io.kroxylicious.proxy.filter.AlterClientQuotasResponseFilter;
+import io.kroxylicious.proxy.filter.AlterConfigsRequestFilter;
+import io.kroxylicious.proxy.filter.AlterConfigsResponseFilter;
+import io.kroxylicious.proxy.filter.AlterPartitionReassignmentsRequestFilter;
+import io.kroxylicious.proxy.filter.AlterPartitionReassignmentsResponseFilter;
+import io.kroxylicious.proxy.filter.AlterPartitionRequestFilter;
+import io.kroxylicious.proxy.filter.AlterPartitionResponseFilter;
+import io.kroxylicious.proxy.filter.AlterReplicaLogDirsRequestFilter;
+import io.kroxylicious.proxy.filter.AlterReplicaLogDirsResponseFilter;
+import io.kroxylicious.proxy.filter.AlterUserScramCredentialsRequestFilter;
+import io.kroxylicious.proxy.filter.AlterUserScramCredentialsResponseFilter;
+import io.kroxylicious.proxy.filter.ApiVersionsRequestFilter;
+import io.kroxylicious.proxy.filter.ApiVersionsResponseFilter;
+import io.kroxylicious.proxy.filter.BeginQuorumEpochRequestFilter;
+import io.kroxylicious.proxy.filter.BeginQuorumEpochResponseFilter;
+import io.kroxylicious.proxy.filter.BrokerHeartbeatRequestFilter;
+import io.kroxylicious.proxy.filter.BrokerHeartbeatResponseFilter;
+import io.kroxylicious.proxy.filter.BrokerRegistrationRequestFilter;
+import io.kroxylicious.proxy.filter.BrokerRegistrationResponseFilter;
+import io.kroxylicious.proxy.filter.CreateAclsRequestFilter;
+import io.kroxylicious.proxy.filter.CreateAclsResponseFilter;
+import io.kroxylicious.proxy.filter.CreateDelegationTokenRequestFilter;
+import io.kroxylicious.proxy.filter.CreateDelegationTokenResponseFilter;
+import io.kroxylicious.proxy.filter.CreatePartitionsRequestFilter;
+import io.kroxylicious.proxy.filter.CreatePartitionsResponseFilter;
+import io.kroxylicious.proxy.filter.CreateTopicsRequestFilter;
+import io.kroxylicious.proxy.filter.CreateTopicsResponseFilter;
+import io.kroxylicious.proxy.filter.DeleteAclsRequestFilter;
+import io.kroxylicious.proxy.filter.DeleteAclsResponseFilter;
+import io.kroxylicious.proxy.filter.DeleteGroupsRequestFilter;
+import io.kroxylicious.proxy.filter.DeleteGroupsResponseFilter;
+import io.kroxylicious.proxy.filter.DeleteRecordsRequestFilter;
+import io.kroxylicious.proxy.filter.DeleteRecordsResponseFilter;
+import io.kroxylicious.proxy.filter.DeleteTopicsRequestFilter;
+import io.kroxylicious.proxy.filter.DeleteTopicsResponseFilter;
+import io.kroxylicious.proxy.filter.DescribeAclsRequestFilter;
+import io.kroxylicious.proxy.filter.DescribeAclsResponseFilter;
+import io.kroxylicious.proxy.filter.DescribeClientQuotasRequestFilter;
+import io.kroxylicious.proxy.filter.DescribeClientQuotasResponseFilter;
+import io.kroxylicious.proxy.filter.DescribeClusterRequestFilter;
+import io.kroxylicious.proxy.filter.DescribeClusterResponseFilter;
+import io.kroxylicious.proxy.filter.DescribeConfigsRequestFilter;
+import io.kroxylicious.proxy.filter.DescribeConfigsResponseFilter;
+import io.kroxylicious.proxy.filter.DescribeDelegationTokenRequestFilter;
+import io.kroxylicious.proxy.filter.DescribeDelegationTokenResponseFilter;
+import io.kroxylicious.proxy.filter.DescribeGroupsRequestFilter;
+import io.kroxylicious.proxy.filter.DescribeGroupsResponseFilter;
+import io.kroxylicious.proxy.filter.DescribeLogDirsRequestFilter;
+import io.kroxylicious.proxy.filter.DescribeLogDirsResponseFilter;
+import io.kroxylicious.proxy.filter.DescribeProducersRequestFilter;
+import io.kroxylicious.proxy.filter.DescribeProducersResponseFilter;
+import io.kroxylicious.proxy.filter.DescribeQuorumRequestFilter;
+import io.kroxylicious.proxy.filter.DescribeQuorumResponseFilter;
+import io.kroxylicious.proxy.filter.DescribeTransactionsRequestFilter;
+import io.kroxylicious.proxy.filter.DescribeTransactionsResponseFilter;
+import io.kroxylicious.proxy.filter.DescribeUserScramCredentialsRequestFilter;
+import io.kroxylicious.proxy.filter.DescribeUserScramCredentialsResponseFilter;
+import io.kroxylicious.proxy.filter.ElectLeadersRequestFilter;
+import io.kroxylicious.proxy.filter.ElectLeadersResponseFilter;
+import io.kroxylicious.proxy.filter.EndQuorumEpochRequestFilter;
+import io.kroxylicious.proxy.filter.EndQuorumEpochResponseFilter;
+import io.kroxylicious.proxy.filter.EndTxnRequestFilter;
+import io.kroxylicious.proxy.filter.EndTxnResponseFilter;
+import io.kroxylicious.proxy.filter.EnvelopeRequestFilter;
+import io.kroxylicious.proxy.filter.EnvelopeResponseFilter;
+import io.kroxylicious.proxy.filter.ExpireDelegationTokenRequestFilter;
+import io.kroxylicious.proxy.filter.ExpireDelegationTokenResponseFilter;
+import io.kroxylicious.proxy.filter.FetchRequestFilter;
+import io.kroxylicious.proxy.filter.FetchResponseFilter;
+import io.kroxylicious.proxy.filter.FetchSnapshotRequestFilter;
+import io.kroxylicious.proxy.filter.FetchSnapshotResponseFilter;
+import io.kroxylicious.proxy.filter.Filter;
+import io.kroxylicious.proxy.filter.FilterContext;
+import io.kroxylicious.proxy.filter.FindCoordinatorRequestFilter;
+import io.kroxylicious.proxy.filter.FindCoordinatorResponseFilter;
+import io.kroxylicious.proxy.filter.HeartbeatRequestFilter;
+import io.kroxylicious.proxy.filter.HeartbeatResponseFilter;
+import io.kroxylicious.proxy.filter.IncrementalAlterConfigsRequestFilter;
+import io.kroxylicious.proxy.filter.IncrementalAlterConfigsResponseFilter;
+import io.kroxylicious.proxy.filter.InitProducerIdRequestFilter;
+import io.kroxylicious.proxy.filter.InitProducerIdResponseFilter;
+import io.kroxylicious.proxy.filter.JoinGroupRequestFilter;
+import io.kroxylicious.proxy.filter.JoinGroupResponseFilter;
+import io.kroxylicious.proxy.filter.LeaveGroupRequestFilter;
+import io.kroxylicious.proxy.filter.LeaveGroupResponseFilter;
+import io.kroxylicious.proxy.filter.ListGroupsRequestFilter;
+import io.kroxylicious.proxy.filter.ListGroupsResponseFilter;
+import io.kroxylicious.proxy.filter.ListOffsetsRequestFilter;
+import io.kroxylicious.proxy.filter.ListOffsetsResponseFilter;
+import io.kroxylicious.proxy.filter.ListPartitionReassignmentsRequestFilter;
+import io.kroxylicious.proxy.filter.ListPartitionReassignmentsResponseFilter;
+import io.kroxylicious.proxy.filter.ListTransactionsRequestFilter;
+import io.kroxylicious.proxy.filter.ListTransactionsResponseFilter;
+import io.kroxylicious.proxy.filter.MetadataRequestFilter;
+import io.kroxylicious.proxy.filter.MetadataResponseFilter;
+import io.kroxylicious.proxy.filter.OffsetCommitRequestFilter;
+import io.kroxylicious.proxy.filter.OffsetCommitResponseFilter;
+import io.kroxylicious.proxy.filter.OffsetDeleteRequestFilter;
+import io.kroxylicious.proxy.filter.OffsetDeleteResponseFilter;
+import io.kroxylicious.proxy.filter.OffsetFetchRequestFilter;
+import io.kroxylicious.proxy.filter.OffsetFetchResponseFilter;
+import io.kroxylicious.proxy.filter.OffsetForLeaderEpochRequestFilter;
+import io.kroxylicious.proxy.filter.OffsetForLeaderEpochResponseFilter;
+import io.kroxylicious.proxy.filter.ProduceRequestFilter;
+import io.kroxylicious.proxy.filter.ProduceResponseFilter;
+import io.kroxylicious.proxy.filter.RenewDelegationTokenRequestFilter;
+import io.kroxylicious.proxy.filter.RenewDelegationTokenResponseFilter;
+import io.kroxylicious.proxy.filter.RequestFilterResult;
+import io.kroxylicious.proxy.filter.ResponseFilterResult;
+import io.kroxylicious.proxy.filter.SaslAuthenticateRequestFilter;
+import io.kroxylicious.proxy.filter.SaslAuthenticateResponseFilter;
+import io.kroxylicious.proxy.filter.SaslHandshakeRequestFilter;
+import io.kroxylicious.proxy.filter.SaslHandshakeResponseFilter;
+import io.kroxylicious.proxy.filter.SyncGroupRequestFilter;
+import io.kroxylicious.proxy.filter.SyncGroupResponseFilter;
+import io.kroxylicious.proxy.filter.TxnOffsetCommitRequestFilter;
+import io.kroxylicious.proxy.filter.TxnOffsetCommitResponseFilter;
+import io.kroxylicious.proxy.filter.UnregisterBrokerRequestFilter;
+import io.kroxylicious.proxy.filter.UnregisterBrokerResponseFilter;
+import io.kroxylicious.proxy.filter.UpdateFeaturesRequestFilter;
+import io.kroxylicious.proxy.filter.UpdateFeaturesResponseFilter;
+import io.kroxylicious.proxy.filter.VoteRequestFilter;
+import io.kroxylicious.proxy.filter.VoteResponseFilter;
+import io.kroxylicious.proxy.filter.WriteTxnMarkersRequestFilter;
+import io.kroxylicious.proxy.filter.WriteTxnMarkersResponseFilter;
+import io.kroxylicious.proxy.internal.filter.AddOffsetsToTxnRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.AddOffsetsToTxnResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.AddPartitionsToTxnRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.AddPartitionsToTxnResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.AllocateProducerIdsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.AllocateProducerIdsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.AlterClientQuotasRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.AlterClientQuotasResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.AlterConfigsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.AlterConfigsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.AlterPartitionReassignmentsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.AlterPartitionReassignmentsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.AlterPartitionRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.AlterPartitionResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.AlterReplicaLogDirsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.AlterReplicaLogDirsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.AlterUserScramCredentialsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.AlterUserScramCredentialsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.ApiVersionsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.ApiVersionsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.BeginQuorumEpochRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.BeginQuorumEpochResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.BrokerHeartbeatRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.BrokerHeartbeatResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.BrokerRegistrationRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.BrokerRegistrationResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.CreateAclsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.CreateAclsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.CreateDelegationTokenRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.CreateDelegationTokenResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.CreatePartitionsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.CreatePartitionsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.CreateTopicsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.CreateTopicsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DeleteAclsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DeleteAclsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DeleteGroupsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DeleteGroupsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DeleteRecordsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DeleteRecordsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DeleteTopicsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DeleteTopicsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DescribeAclsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DescribeAclsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DescribeClientQuotasRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DescribeClientQuotasResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DescribeClusterRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DescribeClusterResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DescribeConfigsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DescribeConfigsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DescribeDelegationTokenRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DescribeDelegationTokenResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DescribeGroupsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DescribeGroupsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DescribeLogDirsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DescribeLogDirsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DescribeProducersRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DescribeProducersResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DescribeQuorumRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DescribeQuorumResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DescribeTransactionsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DescribeTransactionsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DescribeUserScramCredentialsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.DescribeUserScramCredentialsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.ElectLeadersRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.ElectLeadersResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.EndQuorumEpochRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.EndQuorumEpochResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.EndTxnRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.EndTxnResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.EnvelopeRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.EnvelopeResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.ExpireDelegationTokenRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.ExpireDelegationTokenResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.FetchRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.FetchResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.FetchSnapshotRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.FetchSnapshotResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.FilterInvoker;
+import io.kroxylicious.proxy.internal.filter.FilterInvokers;
+import io.kroxylicious.proxy.internal.filter.FindCoordinatorRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.FindCoordinatorResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.HeartbeatRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.HeartbeatResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.IncrementalAlterConfigsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.IncrementalAlterConfigsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.InitProducerIdRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.InitProducerIdResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.JoinGroupRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.JoinGroupResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.LeaveGroupRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.LeaveGroupResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.ListGroupsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.ListGroupsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.ListOffsetsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.ListOffsetsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.ListPartitionReassignmentsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.ListPartitionReassignmentsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.ListTransactionsRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.ListTransactionsResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.MetadataRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.MetadataResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.OffsetCommitRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.OffsetCommitResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.OffsetDeleteRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.OffsetDeleteResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.OffsetFetchRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.OffsetFetchResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.OffsetForLeaderEpochRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.OffsetForLeaderEpochResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.ProduceRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.ProduceResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.RenewDelegationTokenRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.RenewDelegationTokenResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.SaslAuthenticateRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.SaslAuthenticateResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.SaslHandshakeRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.SaslHandshakeResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.SyncGroupRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.SyncGroupResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.TxnOffsetCommitRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.TxnOffsetCommitResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.UnregisterBrokerRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.UnregisterBrokerResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.UpdateFeaturesRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.UpdateFeaturesResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.VoteRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.VoteResponseFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.WriteTxnMarkersRequestFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.WriteTxnMarkersResponseFilterInvoker;
 
 /**
  * Invoker for Filters that implement any number of Specific Message interfaces (for
