+++ kroxylicious-microbenchmarks/src/main/java/io/kroxylicious/microbenchmarks/InvokerScalabilityBenchmark.java
@@ -19,13 +19,13 @@ import org.openjdk.jmh.annotations.Threads;
 import org.openjdk.jmh.annotations.Warmup;
 import org.openjdk.jmh.infra.Blackhole;
 
+import io.kroxylicious.benchmarking.jmh.ArrayFilterInvoker;
+import io.kroxylicious.benchmarking.jmh.SpecificFilterInvoker;
 import io.kroxylicious.microbenchmarks.filters.TwoInterfaceFilter0;
 import io.kroxylicious.microbenchmarks.filters.TwoInterfaceFilter1;
-import io.kroxylicious.proxy.filter.ArrayFilterInvoker;
 import io.kroxylicious.proxy.filter.Filter;
-import io.kroxylicious.proxy.filter.FilterInvoker;
-import io.kroxylicious.proxy.filter.FilterInvokers;
-import io.kroxylicious.proxy.filter.SpecificFilterInvoker;
+import io.kroxylicious.proxy.internal.filter.FilterInvoker;
+import io.kroxylicious.proxy.internal.filter.FilterInvokers;
 
 // try hard to make shouldHandleXYZ to observe different receivers concrete types, saving unrolling to bias a specific call-site to a specific concrete type
 @Fork(value = 2, jvmArgsAppend = "-XX:LoopUnrollLimit=1")
