+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/internal/filter/FilterAndInvoker.java
@@ -4,9 +4,11 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.proxy.filter;
+package io.kroxylicious.proxy.internal.filter;
 
 import java.util.List;
+
+import io.kroxylicious.proxy.filter.Filter;
 
 import static java.util.Objects.requireNonNull;
