+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/internal/filter/HandleNothingFilterInvoker.java
@@ -4,7 +4,7 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.proxy.filter;
+package io.kroxylicious.proxy.internal.filter;
 
 import java.util.concurrent.CompletionStage;
 
@@ -13,8 +13,12 @@ import org.apache.kafka.common.message.ResponseHeaderData;
 import org.apache.kafka.common.protocol.ApiKeys;
 import org.apache.kafka.common.protocol.ApiMessage;
 
-record HandleNothingFilterInvoker() implements FilterInvoker {
-    static final FilterInvoker INSTANCE = new HandleNothingFilterInvoker();
+import io.kroxylicious.proxy.filter.FilterContext;
+import io.kroxylicious.proxy.filter.RequestFilterResult;
+import io.kroxylicious.proxy.filter.ResponseFilterResult;
+
+public record HandleNothingFilterInvoker() implements FilterInvoker {
+    public static final FilterInvoker INSTANCE = new HandleNothingFilterInvoker();
 
     @Override
     public CompletionStage<RequestFilterResult> onRequest(ApiKeys apiKey, short apiVersion, RequestHeaderData header, ApiMessage body, FilterContext filterContext) {
