+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/internal/filter/ResponseFilterInvoker.java
@@ -4,7 +4,7 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.proxy.filter;
+package io.kroxylicious.proxy.internal.filter;
 
 import java.util.concurrent.CompletionStage;
 
@@ -12,7 +12,11 @@ import org.apache.kafka.common.message.ResponseHeaderData;
 import org.apache.kafka.common.protocol.ApiKeys;
 import org.apache.kafka.common.protocol.ApiMessage;
 
-record ResponseFilterInvoker(ResponseFilter filter) implements FilterInvoker {
+import io.kroxylicious.proxy.filter.FilterContext;
+import io.kroxylicious.proxy.filter.ResponseFilter;
+import io.kroxylicious.proxy.filter.ResponseFilterResult;
+
+public record ResponseFilterInvoker(ResponseFilter filter) implements FilterInvoker {
 
     @Override
     public CompletionStage<ResponseFilterResult> onResponse(ApiKeys apiKey, short apiVersion, ResponseHeaderData header, ApiMessage body, FilterContext filterContext) {
