+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/internal/filter/SafeInvoker.java
@@ -4,7 +4,7 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.proxy.filter;
+package io.kroxylicious.proxy.internal.filter;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
@@ -16,6 +16,10 @@ import org.apache.kafka.common.protocol.ApiMessage;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import io.kroxylicious.proxy.filter.FilterContext;
+import io.kroxylicious.proxy.filter.RequestFilterResult;
+import io.kroxylicious.proxy.filter.ResponseFilterResult;
+
 /**
  * Wraps a delegate invoker so that onRequest and onResponse can be safely called even if this
  * Invoker does not want to handle this message, in this case the message will be forwarded without
@@ -23,7 +27,7 @@ import org.slf4j.LoggerFactory;
  *
  * @param invoker the delegate
  */
-record SafeInvoker(FilterInvoker invoker) implements FilterInvoker {
+public record SafeInvoker(FilterInvoker invoker) implements FilterInvoker {
 
     private static final Logger logger = LoggerFactory.getLogger(SafeInvoker.class);
