+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/internal/filter/impl/EagerMetadataLearner.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.kroxylicious.proxy.internal.filter;
+package io.kroxylicious.proxy.internal.filter.impl;
 
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
