+++ kroxylicious-runtime/src/test/java/io/kroxylicious/proxy/internal/filter/ApiVersionsDowngradeFilterTest.java
@@ -22,6 +22,7 @@ import io.kroxylicious.proxy.frame.DecodedRequestFrame;
 import io.kroxylicious.proxy.frame.DecodedResponseFrame;
 import io.kroxylicious.proxy.internal.ApiVersionsServiceImpl;
 import io.kroxylicious.proxy.internal.FilterHarness;
+import io.kroxylicious.proxy.internal.filter.impl.ApiVersionsDowngradeFilter;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
