+++ kroxylicious-runtime/src/test/java/io/kroxylicious/proxy/internal/filter/impl/TopicNameCacheFilterTest.java
@@ -4,7 +4,7 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.proxy.internal.filter;
+package io.kroxylicious.proxy.internal.filter.impl;
 
 import java.time.Duration;
 import java.time.temporal.ChronoUnit;
